+++ apps/website/src/components/dashboard/app-sidebar.tsx
@@ -170,7 +170,7 @@ export function AppSidebar() {
               <DropdownMenuTrigger asChild>
                 <SidebarMenuButton size="lg">
                   <Avatar className="size-8">
-                    <AvatarImage src="/placeholder-user.jpg" />
+                    <AvatarImage src="/soundkit-default-avatar.svg" />
                     <AvatarFallback>JD</AvatarFallback>
                   </Avatar>
                   <div className="flex flex-col gap-0.5 leading-none">